requirements.txt
streamlit
pandas
numpy
matplotlib
Pillow
python-docx
PyPDF2
together
openpyxl